Code style fixes and cleanups
git-svn-id: svn://scribus.net/trunk/Scribus@23534 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/text/storytext.cpp
@@ -2119,7 +2119,7 @@ void StoryText::invalidateAll()
 
 void StoryText::invalidate(int firstItem, int endItem)
 {
-	for (int i=firstItem; i < endItem; ++i)
+	for (int i = firstItem; i < endItem; ++i)
 	{
 		ParagraphStyle* par = item(i)->parstyle;
 		if (par)
@@ -2128,7 +2128,6 @@ void StoryText::invalidate(int firstItem, int endItem)
 	if (!signalsBlocked())
 		emit changed(firstItem, endItem);
 }
-
 
 // physical view
 
@@ -2144,52 +2143,17 @@ void StoryText::validate()
 }
 */
 
-
-
-ScText*  StoryText::item(uint itm)
+ScText*  StoryText::item(int itm)
 {
-	assert( static_cast<int>(itm) < length() );
+	assert( itm < length() );
 	return const_cast<StoryText *>(this)->d->at(itm);
 }
 
-
-const ScText*  StoryText::item(uint itm) const
+const ScText*  StoryText::item(int itm) const
 {
-	assert( static_cast<int>(itm) < length() );
+	assert( itm < length() );
 	return const_cast<StoryText *>(this)->d->at(itm);
 }
-
-
-//const QString StoryText::itemText(uint itm) const
-//{
-	
-//	assert( static_cast<int>(itm) < length() );
-
-//	return text(itm, 1);
-//}
-
-
-//const CharStyle StoryText::itemStyle(uint itm) const
-//{
-//	assert( static_cast<int>(itm) < length() );
-
-//	return charStyle(itm);
-//}
-	
-
-//int StoryText::startOfItem(uint itm) const
-//{
-//	assert( static_cast<int>(itm) < length() );
-
-//	return itm;
-//}
-
-//int StoryText::endOfItem(uint itm) const
-//{
-//	assert( static_cast<int>(itm) < length() );
-
-//	return itm + 1;
-//}
 
 
 using namespace desaxe;
